#71 feature : Scrap 도메인 관련 Dto builder 패턴 적용
