+++ src/main/java/kimp/scrap/dto/internal/binance/BinanceNoticeDto.java
@@ -1,6 +1,6 @@
-package kimp.scrap.dto.binance;
+package kimp.scrap.dto.internal.binance;
 
-import kimp.notice.dto.notice.ExchangeNoticeDto;
+import kimp.notice.dto.response.ExchangeNoticeDto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
